Refactor local storage helper so there is one for subs banner
